Convert all js to ts
