--- src/app.js
@@ -1,20 +1,0 @@
-const express = require("express");
-const app = express();
-const cors = require("cors");
-const favicon = require("express-favicon");
-const logger = require("morgan");
-
-const mainRouter = require("./routes/mainRouter.js");
-
-// middleware
-app.use(cors());
-app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
-app.use(logger("dev"));
-app.use(express.static("public"));
-app.use(favicon(__dirname + "/public/favicon.ico"));
-
-// routes
-app.use("/api/v1", mainRouter);
-
-module.exports = app;
