+++ src/controllers/mainController.ts
@@ -1,0 +1,9 @@
+import { Request, Response } from "express";
+
+const mainController = (req: Request, res: Response) => {
+  res.json({
+    data: "This is a full stack app!",
+  })
+}
+
+export default mainController
