--- src/server.js
@@ -1,5 +1,0 @@
-const { PORT = 8000 } = process.env;
-const app = require("./app");
-
-const listener = () => console.log(`Listening on Port ${PORT}!`);
-app.listen(PORT, listener);
